deleted reference implementation of binary trie
